requirements.txt
langchain==0.1.17
langchain-openai==0.1.4
python-dotenv==1.0.1
sentence-transformers==2.7.0
qdrant-client==1.9.0
pandas==2.2.2
